feat(a11y): add toggle animation menu button
- add menu list component
- refactor state management

[#170446561]
