+++ src/components/compounds/Handicap/HandicapMenuList/handicapMenuList.styles.scss
@@ -1,0 +1,16 @@
+.handicapMenu {
+  height: fit-content;
+  width: fit-content;
+  margin-left: 10px;
+}
+
+.handicapMenuClosed {
+  width: 0;
+  padding: 0;
+  visibility: hidden;
+}
+
+.animatedIcon {
+  background-size: cover;
+  background-image: url("../../../../static/images/book.gif");
+}
